improved restore page, code cleanup
