+++ fyne/assets/icons_test.go
@@ -9,7 +9,7 @@ func TestGetIcons(t *testing.T) {
 	t.Run("given icons that exist in the assets directory, maps each one to a fyne static resource", func(t *testing.T) {
 		iconNames := []string{
 			AccountsIcon, HistoryIcon, OverviewIcon, ReceiveIcon, SendIcon, StakingIcon}
-		icons, err := GetIcons(iconNames...)
+		icons, err := Get(iconNames...)
 		if err != nil {
 			t.Errorf("getIcons returned an error for icons in the assets directory: %v", err)
 			t.FailNow()
@@ -26,7 +26,7 @@ func TestGetIcons(t *testing.T) {
 	})
 	t.Run("returns an error for icons that do not exist in the assets directory", func(t *testing.T) {
 		iconNames := []string{"not_existing.png", "deleted.png"}
-		_, err := GetIcons(iconNames...)
+		_, err := Get(iconNames...)
 		if err == nil {
 			t.Errorf("expected an error for icon names %v, got nil", iconNames)
 		}
