+++ fyne/pages/spendingpassword.go
@@ -1,0 +1,189 @@
+package pages
+
+import (
+	"fmt"
+	"image/color"
+	"log"
+
+	"github.com/raedahgroup/godcr/fyne/layouts"
+
+	"fyne.io/fyne"
+	"fyne.io/fyne/canvas"
+	"fyne.io/fyne/layout"
+	"fyne.io/fyne/theme"
+	"fyne.io/fyne/widget"
+	"github.com/raedahgroup/dcrlibwallet"
+	"github.com/raedahgroup/godcr/fyne/assets"
+	"github.com/raedahgroup/godcr/fyne/widgets"
+)
+
+func (app *AppInterface) createSpendingPasswordPopup(seed string) {
+	var popup *widget.PopUp
+	popupContent := widget.NewVBox()
+
+	popup = widget.NewModalPopUp(widget.NewHBox(widgets.NewHSpacer(10),
+		widget.NewVBox(widgets.NewVSpacer(10),
+			widget.NewLabelWithStyle("Create a spending password", fyne.TextAlignLeading, fyne.TextStyle{Bold: true}),
+			popupContent, widgets.NewVSpacer(10)),
+		widgets.NewHSpacer(10)),
+		app.Window.Canvas())
+
+	popupContent.Children = []fyne.CanvasObject{app.passwordPopup(popup, seed)}
+	widget.Refresh(popupContent)
+}
+
+func (app *AppInterface) passwordPopup(popup *widget.PopUp, seed string) fyne.CanvasObject {
+	displayError := func(err error) {
+		log.Println("could not generate seed", err.Error())
+		newWindow := fyne.CurrentApp().NewWindow(app.AppDisplayName)
+		newWindow.SetContent(widget.NewVBox(
+			widget.NewLabelWithStyle(fmt.Sprintf("Could not generate seed, %s", err.Error()), fyne.TextAlignCenter, fyne.TextStyle{}),
+			widget.NewHBox(layout.NewSpacer(), widget.NewButton("Close", func() { newWindow.Close() }), layout.NewSpacer())))
+
+		newWindow.CenterOnScreen()
+		newWindow.Show()
+		newWindow.SetFixedSize(true)
+	}
+
+	icons, err := assets.Get(assets.Reveal, assets.Conceal, assets.Loader)
+	if err != nil {
+		return app.DisplayLaunchErrorAndExit(err.Error())
+	}
+
+	errorLabel := canvas.NewText("Password do not match", color.RGBA{255, 0, 0, 255})
+	errorLabel.TextSize = 10
+	errorLabel.Hide()
+
+	password := widget.NewPasswordEntry()
+	password.SetPlaceHolder("Spending Password")
+	confirmPassword := widget.NewPasswordEntry()
+	confirmPassword.SetPlaceHolder("Confirm Spending Password")
+
+	passwordLength := canvas.NewText("0", color.Black)
+	passwordLength.TextSize = 10
+	passwordLength.Alignment = fyne.TextAlignTrailing
+	confirmPasswordLength := canvas.NewText("0", color.Black)
+	confirmPasswordLength.TextSize = 10
+	confirmPasswordLength.Alignment = fyne.TextAlignTrailing
+
+	passwordStrength := widget.NewProgressBar()
+	var createButton *widget.Button
+
+	password.OnChanged = func(val string) {
+		// check if password and confirm password matches only when the user fills confirmPassword textbox
+		if confirmPassword.Text != "" {
+			if confirmPassword.Text != password.Text {
+				errorLabel.Show()
+			} else {
+				errorLabel.Hide()
+				createButton.Enable()
+			}
+		}
+
+		passwordLength.Text = fmt.Sprintf("%d", len(val))
+		canvas.Refresh(passwordLength)
+
+		strength := (dcrlibwallet.ShannonEntropy(val) / 4.0)
+		passwordStrength.SetValue(strength)
+	}
+
+	confirmPassword.OnChanged = func(val string) {
+		confirmPasswordLength.Text = fmt.Sprintf("%d", len(val))
+		canvas.Refresh(confirmPasswordLength)
+		if password.Text != val {
+			errorLabel.Show()
+		} else if password.Text != "" && password.Text == confirmPassword.Text {
+			errorLabel.Hide()
+			createButton.Enable()
+		}
+	}
+
+	cancelLabel := canvas.NewText("Cancel", color.RGBA{41, 112, 255, 255})
+	cancelLabel.TextStyle.Bold = true
+	cancelButton := widgets.NewClickableBox(widget.NewHBox(cancelLabel), func() { popup.Hide() })
+
+	createButton = widget.NewButton("Create", func() {
+		createButton.SetText("")
+		createButton.SetIcon(icons[assets.Loader])
+		createButton.Disable()
+
+		// disable cancel OnTapped function
+		cancelButton.OnTapped = nil
+		cancelLabel.Color = color.RGBA{196, 203, 210, 255}
+		canvas.Refresh(cancelLabel)
+
+		enableCancelButton := func() {
+			cancelButton.OnTapped = nil
+			cancelLabel.Color = color.RGBA{41, 112, 255, 255}
+		}
+
+		var err error
+		if seed == "" {
+			seed, err = dcrlibwallet.GenerateSeed()
+			if err != nil {
+				enableCancelButton()
+				displayError(err)
+				return
+			}
+		}
+
+		err = app.Dcrlw.CreateWallet(password.Text, seed)
+
+		if err != nil {
+			enableCancelButton()
+			displayError(err)
+			log.Println("could not create wallet", err.Error())
+			return
+		}
+
+		popup.Hide()
+		app.Window.SetFixedSize(false)
+		app.Window.SetOnClosed(nil)
+		app.Window.SetContent(app.tabMenu)
+		app.tabMenu.CreateRenderer().ApplyTheme()
+		// apparently tabMenu was initialized in fyne.go line 57,
+		// this sets theme of tabmenu to default black
+		// resetting theme to light theme when tabmenu is in view fixes the font.
+		fyne.CurrentApp().Settings().SetTheme(theme.LightTheme())
+	})
+
+	createButton.Disable()
+
+	var passwordConceal *widgets.ImageButton
+	passwordConceal = widgets.NewImageButton(icons[assets.Reveal], nil, func() {
+		if password.Password {
+			passwordConceal.SetIcon(icons[assets.Conceal])
+			password.Password = false
+		} else {
+			passwordConceal.SetIcon(icons[assets.Reveal])
+			password.Password = true
+		}
+		// reveal texts
+		password.SetText(password.Text)
+	})
+
+	var confirmPasswordConceal *widgets.ImageButton
+	confirmPasswordConceal = widgets.NewImageButton(icons[assets.Reveal], nil, func() {
+		if confirmPassword.Password {
+			confirmPasswordConceal.SetIcon(icons[assets.Conceal])
+			confirmPassword.Password = false
+		} else {
+			confirmPasswordConceal.SetIcon(icons[assets.Reveal])
+			confirmPassword.Password = true
+		}
+		// reveal texts
+		confirmPassword.SetText(confirmPassword.Text)
+	})
+
+	return widget.NewVBox(widgets.NewVSpacer(10),
+		fyne.NewContainerWithLayout(layouts.NewPasswordLayout(fyne.NewSize(312, password.MinSize().Height)), password, passwordConceal),
+		passwordLength,
+		widget.NewHBox(layout.NewSpacer(),
+			fyne.NewContainerWithLayout(
+				layout.NewFixedGridLayout(fyne.NewSize(150, widget.NewLabel("0%").MinSize().Height)), passwordStrength)),
+		fyne.NewContainerWithLayout(layouts.NewPasswordLayout(fyne.NewSize(312, confirmPassword.MinSize().Height)), confirmPassword, confirmPasswordConceal),
+		confirmPasswordLength,
+		widget.NewHBox(layout.NewSpacer(), cancelButton, widgets.NewHSpacer(24), createButton),
+		errorLabel,
+		widgets.NewVSpacer(10))
+}
